pyproject.toml
[project]
name = "google_doc_utils"
version = "0.1.0"
description = "MCP server for Google Docs enabling LLMs to read and edit documents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.100.0",
    "mcp>=1.25.0",
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/extended_google_doc_utils"]

[tool.uv]
dev-dependencies = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "tier_a: Tier A tests - require Google Cloud credentials",
    "tier_b: Tier B tests - use mocks, no credentials required",
]

[tool.coverage.run]
source = ["src/extended_google_doc_utils"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true
fail_under = 70

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []
